requirements.txt
discord.py>=2.3.2
# easyocr>=1.7.1